+++ qtgl/mathutils.hpp
@@ -3,9 +3,9 @@
 namespace qtgl {
 
 struct MathUtils {
-  constexpr static float PI = 3.1415926f;
-  inline static float toRadians(float d) { return d * (PI / 180); }
-  inline static float toDegree(float r) { return r * (180 / PI); }
+  constexpr static double PI = 3.1415926;
+  inline static double toRadians(double d) { return d * (PI / 180); }
+  inline static double toDegree(double r) { return r * (180 / PI); }
 };
 
 }  // namespace qtgl
